Fixed argument

+++ vciicon.sh
@@ -8,9 +8,12 @@ echo "VCiicon Beta by Larsenv."
 echo "Requires ImageMagick for now, and icon to iconTex.png."
 echo "\n"
 mkdir -p "meta"
-convert iconTex.png -resize 122x81\! -gravity center iconTex.png
-convert iconTex.png -gravity center -crop 128x128 iconTex.png
-composite -geometry +3+17 iconTex.png ./resources/iconTex.png ./resources/iconTex.png ./meta/iconTex.tga
-rm iconTex.png
+if [ -n "GBA" ]
+then 
+	convert iconTex.png -resize 122x81\! -gravity center iconTex.png
+	convert iconTex.png -gravity center -crop 128x128 iconTex.png
+	composite -geometry +3+17 iconTex.png ./resources/iconTex-GBA.png ./resources/iconTex-GBA.png ./meta/iconTex.tga
+	rm iconTex.png
+fi
 echo "\n"
 echo "Done."
